extract file-status-update.js from redis-pubsub directory
