+++ src/pubsub-connector.js
@@ -1,7 +1,7 @@
 const podname = process.env.podname;
 const logger = require("./logger");
 const redisPubsub = require("./redis-pubsub");
-const fileStatusUpdate = require("./redis-pubsub/file-status-update");
+const fileStatusUpdate = require("./file-status-update");
 
 module.exports = function postHandler(req, resp) {
   logger.log(`Received from PSC: ${JSON.stringify(req.body, null, 2)}`); // eslint-disable-line
